#  refactor(tsdb/cursors): Add boundSearch to improve FindRange