As title, add boundSearch to reduce some unnecessary checks in the binary search of FindRange.
I add benchmark for the IntegerArray.
```
benchmark                                    old ns/op     new ns/op     delta
BenchmarkIntegerArray_FindRange_100000-8     14166         12611         -10.98%